Create InternetListenerActivity and move to listeners package
