+++ app/src/main/java/com/groupseven/musicmap/screens/verification/VerificationActivity.java
@@ -7,7 +7,7 @@ import android.widget.Button;
 import android.widget.TextView;
 
 import com.groupseven.musicmap.R;
-import com.groupseven.musicmap.SessionAndInternetListenerActivity;
+import com.groupseven.musicmap.listeners.SessionAndInternetListenerActivity;
 import com.groupseven.musicmap.screens.main.HomeActivity;
 import com.groupseven.musicmap.models.Artist;
 import com.groupseven.musicmap.firebase.Session;
